picturas/Tools/generative_expand_ai/requirements.txt
fastapi
uvicorn
pillow
minio
requests
numpy
